api/index.py
import sys
import os
from pathlib import Path

# Add the project root to Python path for imports
project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

# Import the FastAPI app
from backend.main import app

# Vercel expects this for Python functions
def handler(request):
    return app